move search cubit to to search package
