+++ apps/customer/lib/presentation/screens/search/cubit/search_cubit.dart
