+++ apps/customer/lib/presentation/screens/search/cubit/search_state.dart
