+++ apps/customer/lib/presentation/screens/search/widgets/search_bar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import '../../../cubits/search/cubit/search_cubit.dart';
+import '../cubit/search_cubit.dart';
 import '../../home/sections/search/widgets/search_bar_widget.dart';
 
 class SearchBarSection extends StatelessWidget {
